# Optimize polynomial batching computation