Optimizes `PolyEvalWitness::batch` and `PolyEvalWitness::batch_diff_size` 

SHA256 Benchmark, compared against `main`:

```
❯ cargo bench --bench sha256 -- --baseline main
   Compiling nova-snark v0.36.0 (/Users/mzhu/code/Nova)
    Finished `bench` profile [optimized] target(s) in 5.14s
     Running benches/sha256.rs (target/release/deps/sha256-2bc87e6fb709805f)
Gnuplot not found, using plotters backend
NovaProve-Sha256-message-len-64/Prove
                        time:   [38.233 ms 38.350 ms 38.470 ms]
                        change: [+0.2237% +0.8242% +1.5157%] (p = 0.03 < 0.05)
                        Change within noise threshold.

NovaProve-Sha256-message-len-128/Prove
                        time:   [47.714 ms 47.882 ms 48.248 ms]
                        change: [+0.3301% +0.9473% +1.5827%] (p = 0.01 < 0.05)
                        Change within noise threshold.

NovaProve-Sha256-message-len-256/Prove
                        time:   [57.914 ms 58.168 ms 58.357 ms]
                        change: [-1.0506% -0.3042% +0.4057%] (p = 0.45 > 0.05)
                        No change in performance detected.

NovaProve-Sha256-message-len-512/Prove
                        time:   [77.322 ms 77.728 ms 78.063 ms]
                        change: [-1.4300% -0.4703% +0.4200%] (p = 0.36 > 0.05)
                        No change in performance detected.

Benchmarking NovaProve-Sha256-message-len-1024/Prove: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 6.5s or enable flat sampling.
NovaProve-Sha256-message-len-1024/Prove
                        time:   [117.23 ms 117.56 ms 118.06 ms]
                        change: [-1.5108% -0.7987% -0.0734%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 3 outliers among 10 measurements (30.00%)
  1 (10.00%) low mild
  1 (10.00%) high mild
  1 (10.00%) high severe

NovaProve-Sha256-message-len-2048/Prove
                        time:   [196.09 ms 198.28 ms 200.92 ms]
                        change: [-2.7311% -0.9839% +0.8079%] (p = 0.32 > 0.05)
                        No change in performance detected.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) high mild
  1 (10.00%) high severe

NovaProve-Sha256-message-len-4096/Prove
                        time:   [360.62 ms 366.00 ms 372.74 ms]
                        change: [-2.5796% -0.9163% +1.3261%] (p = 0.38 > 0.05)
                        No change in performance detected.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) high mild
  1 (10.00%) high severe

Benchmarking NovaProve-Sha256-message-len-8192/Prove: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 6.9s.
NovaProve-Sha256-message-len-8192/Prove
                        time:   [682.35 ms 727.58 ms 790.57 ms]
                        change: [-5.1123% +2.0321% +13.448%] (p = 0.71 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe

Benchmarking NovaProve-Sha256-message-len-16384/Prove: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 13.2s.
NovaProve-Sha256-message-len-16384/Prove
                        time:   [1.3080 s 1.3599 s 1.4234 s]
                        change: [-15.529% -11.092% -5.9208%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

Benchmarking NovaProve-Sha256-message-len-32768/Prove: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 27.7s.
NovaProve-Sha256-message-len-32768/Prove
                        time:   [2.7410 s 2.8505 s 2.9849 s]
                        change: [-11.713% -6.4613% -1.0214%] (p = 0.05 > 0.05)
                        No change in performance detected.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) high mild
  1 (10.00%) high severe

Benchmarking NovaProve-Sha256-message-len-65536/Prove: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 60.4s.
NovaProve-Sha256-message-len-65536/Prove
                        time:   [5.4715 s 5.7959 s 6.1407 s]
                        change: [-19.865% -13.258% -5.9665%] (p = 0.01 < 0.05)
                        Performance has improved.
```